django_project_boilerplate-master/requirements.txt
argon2-cffi==19.1.0
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
Faker==1.0.7
idna==2.8
oauthlib==3.0.2
Pillow==6.0.0
pycparser==2.19
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
text-unidecode==1.2
urllib3==1.25.3
wincertstore==0.2
